+++ project/src/components/nearby-offer/nearby-offer.jsx
